renamed carsettings.tsx to car.tsx and added an advanced menu
